// src/interfaces/IInternalItemResponse.ts
// Generated by https://quicktype.io
// /api/items/:id

import { Author } from './IAuthor';
import { ICategory } from './IInternalSearchResponse';

export interface Price {
  currency: string;
  amount: number;
  decimals: number;
}

export interface IItemData {
  id: string;
  title: string;
  price: Price;
  picture: string;
  condition: string;
  free_shipping: boolean;
  sold_quantity: number;
  description: string;
  path_from_root: ICategory[];
}

export interface IInternalItemResponse {
  author: Author;
  item: IItemData;
}